Allows to mix programmatic and declarative API defining agentic patterns
