+++ langchain4j-agentic/src/main/java/dev/langchain4j/agentic/supervisor/SupervisorAgentServiceImpl.java
@@ -1,10 +1,19 @@
 package dev.langchain4j.agentic.supervisor;
 
+import static dev.langchain4j.agentic.declarative.DeclarativeUtil.agenticScopeFunction;
+import static dev.langchain4j.agentic.declarative.DeclarativeUtil.buildAgentFeatures;
+import static dev.langchain4j.agentic.declarative.DeclarativeUtil.invokeStatic;
+import static dev.langchain4j.agentic.declarative.DeclarativeUtil.selectMethod;
 import static dev.langchain4j.agentic.internal.AgentUtil.validateAgentClass;
 
+import dev.langchain4j.agentic.declarative.ChatMemoryProviderSupplier;
+import dev.langchain4j.agentic.declarative.ChatModelSupplier;
+import dev.langchain4j.agentic.declarative.Output;
+import dev.langchain4j.agentic.declarative.SupervisorRequest;
 import dev.langchain4j.agentic.internal.AbstractServiceBuilder;
 import dev.langchain4j.agentic.planner.AgenticService;
 import dev.langchain4j.agentic.scope.AgenticScope;
+import dev.langchain4j.memory.ChatMemory;
 import dev.langchain4j.memory.chat.ChatMemoryProvider;
 import dev.langchain4j.model.chat.ChatModel;
 import java.lang.reflect.Method;
@@ -27,6 +36,7 @@ public class SupervisorAgentServiceImpl<T> extends AbstractServiceBuilder<T, Sup
 
     public SupervisorAgentServiceImpl(Class<T> agentServiceClass, Method agenticMethod) {
         super(agentServiceClass, agenticMethod);
+        configureSupervisor(agentServiceClass);
     }
 
     public T build() {
@@ -101,5 +111,36 @@ public class SupervisorAgentServiceImpl<T> extends AbstractServiceBuilder<T, Sup
     @Override
     public String serviceType() {
         return "Supervisor";
+    }
+
+    private void configureSupervisor(Class<T> agentServiceClass) {
+        selectMethod(
+                agentServiceClass,
+                method -> method.isAnnotationPresent(SupervisorRequest.class)
+                        && method.getReturnType() == String.class)
+                .map(m -> agenticScopeFunction(m, String.class))
+                .ifPresent(this::requestGenerator);
+
+        selectMethod(
+                agentServiceClass,
+                method -> method.isAnnotationPresent(ChatModelSupplier.class)
+                        && method.getReturnType() == ChatModel.class
+                        && method.getParameterCount() == 0)
+                .map(method -> (ChatModel) invokeStatic(method))
+                .ifPresent(this::chatModel);
+
+        selectMethod(
+                agentServiceClass,
+                method -> method.isAnnotationPresent(ChatMemoryProviderSupplier.class)
+                        && method.getReturnType() == ChatMemory.class
+                        && method.getParameterCount() == 1)
+                .map(method -> (ChatMemoryProvider) memoryId -> invokeStatic(method, memoryId))
+                .ifPresent(this::chatMemoryProvider);
+
+        selectMethod(agentServiceClass, method -> method.isAnnotationPresent(Output.class))
+                .map(m -> agenticScopeFunction(m, Object.class))
+                .ifPresent(this::output);
+
+        buildAgentFeatures(agentServiceClass, this);
     }
 }
