+++ langchain4j-agentic/src/test/java/dev/langchain4j/agentic/SupervisorAndWorkflowAgentsIT.java
@@ -9,7 +9,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import dev.langchain4j.agentic.Agents.CreativeWriter;
 import dev.langchain4j.agentic.Agents.StyleEditor;
 import dev.langchain4j.agentic.Agents.StyleReviewLoop;
+import dev.langchain4j.agentic.Agents.StyleReviewLoopWithExitCondition;
 import dev.langchain4j.agentic.Agents.StyleScorer;
+import dev.langchain4j.agentic.declarative.SupervisorRequest;
 import dev.langchain4j.agentic.scope.AgentInvocation;
 import dev.langchain4j.agentic.scope.DefaultAgenticScope;
 import dev.langchain4j.agentic.scope.ResultWithAgenticScope;
@@ -30,6 +32,14 @@ public class SupervisorAndWorkflowAgentsIT {
         @Agent
         ResultWithAgenticScope<String> write(
                 @V("topic") String topic, @V("style") String style, @V(SUPERVISOR_CONTEXT_KEY) String businessContext);
+    }
+
+    public interface SupervisorStyledWriterWithRequest extends SupervisorStyledWriter {
+
+        @SupervisorRequest
+        static String request(@V("topic") String topic, @V("style") String style) {
+            return "Write a story about " + topic + " in the style of a " + style;
+        }
     }
 
     @Test
@@ -59,17 +69,14 @@ public class SupervisorAndWorkflowAgentsIT {
                 .outputKey("score")
                 .build();
 
-        StyleReviewLoop styleReviewLoop = AgenticServices.loopBuilder(StyleReviewLoop.class)
+        StyleReviewLoopWithExitCondition styleReviewLoop = AgenticServices.loopBuilder(StyleReviewLoopWithExitCondition.class)
                 .subAgents(styleScorer, styleEditor)
                 .outputKey("story")
                 .maxIterations(5)
-                .exitCondition(agenticScope -> agenticScope.readState("score", 0.0) >= 0.8)
                 .build();
 
-        SupervisorStyledWriter styledWriter = AgenticServices.supervisorBuilder(SupervisorStyledWriter.class)
+        SupervisorStyledWriter styledWriter = AgenticServices.supervisorBuilder(SupervisorStyledWriterWithRequest.class)
                 .chatModel(plannerModel())
-                .requestGenerator(agenticScope -> "Write a story about " + agenticScope.readState("topic")
-                        + " in the style of a " + agenticScope.readState("style"))
                 .responseStrategy(SupervisorResponseStrategy.LAST)
                 .subAgents(creativeWriter, styleReviewLoop)
                 .maxAgentsInvocations(5)
